Fix AttributeError in keystone functional tests
OpenStack allows omitting names for services. This patch adds a check
to ensure we do not call startswith on None values.

+++ shade/tests/functional/test_services.py
@@ -47,7 +47,8 @@ class TestServices(base.TestCase):
     def _cleanup_services(self):
         exception_list = list()
         for s in self.operator_cloud.list_services():
-            if s['name'].startswith(self.new_service_name):
+            if s['name'] is not None and \
+                    s['name'].startswith(self.new_service_name):
                 try:
                     self.operator_cloud.delete_service(name_or_id=s['id'])
                 except Exception as e:
